+++ test/controllers/consignors/ConsignorNameControllerSpec.scala
@@ -16,10 +16,10 @@
 
 package controllers.consignors
 
-import controllers.{routes => baseRoutes}
 import base.SpecBase
-import forms.consignors.ConsignorsIdentityFormProvider
-import models.{ConsignorsIdentity, NormalMode}
+import controllers.{routes => baseRoutes}
+import forms.consignors.ConsignorNameFormProvider
+import models.NormalMode
 import org.mockito.ArgumentMatchers.{any, eq => eqTo}
 import org.mockito.Mockito.{times, verify, when}
 import org.scalatestplus.mockito.MockitoSugar
@@ -28,30 +28,30 @@ import play.api.inject.bind
 import play.api.test.FakeRequest
 import play.api.test.Helpers._
 import repositories.SessionRepository
-import views.html.consignors.ConsignorsIdentityView
+import views.html.consignors.ConsignorNameView
 
 import scala.concurrent.Future
 
-class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
+class ConsignorNameControllerSpec extends SpecBase with MockitoSugar {
 
-  lazy val consignorsIdentityRoute =
-    routes.ConsignorsIdentityController.onPageLoad(NormalMode, lrn, index).url
-
-  val formProvider = new ConsignorsIdentityFormProvider()
+  val formProvider = new ConsignorNameFormProvider()
   val form = formProvider()
 
-  "ConsignorsIdentity Controller" - {
+  lazy val consignorNameRoute =
+    routes.ConsignorNameController.onPageLoad(NormalMode, lrn, index).url
+
+  "ConsignorName Controller" - {
 
     "must return OK and the correct view for a GET" in {
 
       val application = applicationBuilder(userAnswers = Some(emptyUserAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, consignorsIdentityRoute)
+        val request = FakeRequest(GET, consignorNameRoute)
 
         val result = route(application, request).value
 
-        val view = application.injector.instanceOf[ConsignorsIdentityView]
+        val view = application.injector.instanceOf[ConsignorNameView]
 
         status(result) mustEqual OK
         contentAsString(result) mustEqual view(form, NormalMode, lrn, index)(
@@ -63,27 +63,22 @@ class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
 
     "must populate the view correctly on a GET when the question has previously been answered" in {
 
-      val userAnswers = emptyUserAnswers
-        .set(consignors.ConsignorsIdentityPage(index), ConsignorsIdentity.values.head)
-        .success
-        .value
+      val userAnswers = emptyUserAnswers.set(consignors.ConsignorNamePage(index), "answer").success.value
 
       val application = applicationBuilder(userAnswers = Some(userAnswers)).build()
 
       running(application) {
-        val request = FakeRequest(GET, consignorsIdentityRoute)
+        val request = FakeRequest(GET, consignorNameRoute)
 
-        val view = application.injector.instanceOf[ConsignorsIdentityView]
+        val view = application.injector.instanceOf[ConsignorNameView]
 
         val result = route(application, request).value
 
         status(result) mustEqual OK
-        contentAsString(result) mustEqual view(
-          form.fill(ConsignorsIdentity.values.head),
-          NormalMode,
-          lrn,
-          index
-        )(request, messages(application)).toString
+        contentAsString(result) mustEqual view(form.fill("answer"), NormalMode, lrn, index)(
+          request,
+          messages(application)
+        ).toString
       }
     }
 
@@ -100,17 +95,14 @@ class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, consignorsIdentityRoute)
-            .withFormUrlEncodedBody(("value", ConsignorsIdentity.values.head.toString))
+          FakeRequest(POST, consignorNameRoute)
+            .withFormUrlEncodedBody(("value", "answer"))
 
         val result = route(application, request).value
-        val expectedAnswers = emptyUserAnswers
-          .set(consignors.ConsignorsIdentityPage(index), ConsignorsIdentity.values.head)
-          .success
-          .value
+        val expectedAnswers = emptyUserAnswers.set(consignors.ConsignorNamePage(index), "answer").success.value
 
         status(result) mustEqual SEE_OTHER
-        redirectLocation(result).value mustEqual consignors.ConsignorsIdentityPage(index)
+        redirectLocation(result).value mustEqual consignors.ConsignorNamePage(index)
           .navigate(NormalMode, expectedAnswers)
           .url
         verify(mockSessionRepository, times(1)).set(eqTo(expectedAnswers))
@@ -123,12 +115,12 @@ class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
 
       running(application) {
         val request =
-          FakeRequest(POST, consignorsIdentityRoute)
-            .withFormUrlEncodedBody(("value", "invalid value"))
+          FakeRequest(POST, consignorNameRoute)
+            .withFormUrlEncodedBody(("value", ""))
 
-        val boundForm = form.bind(Map("value" -> "invalid value"))
+        val boundForm = form.bind(Map("value" -> ""))
 
-        val view = application.injector.instanceOf[ConsignorsIdentityView]
+        val view = application.injector.instanceOf[ConsignorNameView]
 
         val result = route(application, request).value
 
@@ -145,7 +137,7 @@ class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
       val application = applicationBuilder(userAnswers = None).build()
 
       running(application) {
-        val request = FakeRequest(GET, consignorsIdentityRoute)
+        val request = FakeRequest(GET, consignorNameRoute)
 
         val result = route(application, request).value
 
@@ -154,19 +146,18 @@ class ConsignorsIdentityControllerSpec extends SpecBase with MockitoSugar {
       }
     }
 
-    "redirect to Journey Recovery for a POST if no existing data is found" in {
+    "must redirect to Journey Recovery for a POST if no existing data is found" in {
 
       val application = applicationBuilder(userAnswers = None).build()
 
       running(application) {
         val request =
-          FakeRequest(POST, consignorsIdentityRoute)
-            .withFormUrlEncodedBody(("value", ConsignorsIdentity.values.head.toString))
+          FakeRequest(POST, consignorNameRoute)
+            .withFormUrlEncodedBody(("value", "answer"))
 
         val result = route(application, request).value
 
         status(result) mustEqual SEE_OTHER
-
         redirectLocation(result).value mustEqual baseRoutes.JourneyRecoveryController.onPageLoad().url
       }
     }
